+++ lib/ts/recipe/thirdparty/providers/bitbucket.ts
@@ -80,12 +80,12 @@ export default function Bitbucket(input: ProviderInput): TypeProvider {
                 logDebugMessage(
                     `Received response with status ${
                         userInfoFromAccessToken.status
-                    } and body ${await userInfoFromAccessToken.response.clone().text()}`
+                    } and body ${await userInfoFromAccessToken.rawResponse.text()}`
                 );
                 throw new Error(
                     `Received response with status ${
                         userInfoFromAccessToken.status
-                    } and body ${await userInfoFromAccessToken.response.clone().text()}`
+                    } and body ${await userInfoFromAccessToken.rawResponse.text()}`
                 );
             }
             rawUserInfoFromProvider.fromUserInfoAPI = userInfoFromAccessToken.response;
@@ -100,12 +100,12 @@ export default function Bitbucket(input: ProviderInput): TypeProvider {
                 logDebugMessage(
                     `Received response with status ${
                         userInfoFromEmail.status
-                    } and body ${await userInfoFromEmail.response.clone().text()}`
+                    } and body ${await userInfoFromEmail.rawResponse.text()}`
                 );
                 throw new Error(
                     `Received response with status ${
                         userInfoFromEmail.status
-                    } and body ${await userInfoFromEmail.response.clone().text()}`
+                    } and body ${await userInfoFromEmail.rawResponse.text()}`
                 );
             }