+++ CHANGELOG.md
@@ -5,11 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
-## [16.2.2] - 2023-10-09
+## [16.3.0] - 2023-10-10
+
+### Added
+
+-   Adds support for configuring multiple frontend domains to be used with the same backend
+-   Added a new `origin` property to `appInfo`, this can be configured to be a function which allows you to conditionally return the value of the frontend domain. This property will replace `websiteDomain` in a future release of `supertokens-node`
+-   `websiteDomain` inside `appInfo` is now optional. Using `origin` is recommended over using `websiteDomain`. This is not a breaking change and using `websiteDomain` will continue to work
+-   Added a "custom" framework you can use in framework normally not supported by our SDK
+-   Added a next13 app router compatible request handler.
+
+### Fixed
+
+-   Fixed an issue where calling signinup for thirdparty recipe would result in a "clone is not a function" error
 
 ### Changes
 
 -   Using built-in fetch whenever available instead of cross-fetch
+-   Improved edge-function support
 
 ## [16.2.1] - 2023-10-06
 
@@ -100,6 +113,10 @@ Because of account linking we've introduced a new Primary user concept (see abov
         -   returns new `recipeUserId` prop in the `status: OK` case
     -   `signUp`:
         -   returns new `recipeUserId` prop in the `status: OK` case
+    -   `resetPasswordUsingToken`:
+        -   removed from the recipe interface, making it no longer overrideable (directly)
+        -   the related function in the index files now call `consumePasswordResetToken` and `updateEmailOrPassword`
+        -   any necessary behaviour changes can be achieved by overriding those two function instead
     -   `signInPOST`:
         -   can return status `SIGN_IN_NOT_ALLOWED`
     -   `signUpPOST`:
@@ -160,6 +177,10 @@ Because of account linking we've introduced a new Primary user concept (see abov
     -   `updateEmailOrPassword` :
         -   now takes `recipeUserId` instead of `userId`
         -   can return the new `EMAIL_CHANGE_NOT_ALLOWED_ERROR` status
+    -   `resetPasswordUsingToken`:
+        -   removed from the recipe interface, making it no longer overrideable (directly)
+        -   the related function in the index files now call `consumePasswordResetToken` and `updateEmailOrPassword`
+        -   any necessary behaviour changes can be achieved by overriding those two function instead
     -   added an overrideable `createNewEmailPasswordRecipeUser` function that is called during sign up and in the “invitation link” flow
     -   `emailPasswordSignIn`:
         -   returns new `recipeUserId` prop in the `status: OK` case